# Add HTTP:// as default prefix for urls