Old behavior
$ benchstats www.github.com
017/03/26 01:11:36 request failed: Get www.github.com: unsupported protocol scheme ""
exit status 1

New behavior
$ benchstats www.github.com
200 OK
Average request time: 0.589077682s
DNS Lookup: 0.013210684s
TCP Connections: 0.135449342s
Server Procesing: 0.157549961s
Server Tranfer: 0.000324132s